permission checking (finally)
improved .roll and added usage indactors to .help [command]

+++ commands/purge.py
@@ -2,6 +2,7 @@ import discord
 from discord.ext import commands
 from datetime import date
 from datetime import datetime
+from Main import NoPermsEmbed
 
 class Admin(commands.Cog):
     def __init__(self, client):
@@ -9,28 +10,32 @@ class Admin(commands.Cog):
 
     @commands.command()
     async def purge(self, ctx, msgCount):
-        today = date.today()
-        hour = datetime.now()
-        hourString = hour.strftime("%H:%M:%S")
+        if ctx.message.author.guild_permissions.manage_channels:
+            today = date.today()
+            hour = datetime.now()
+            hourString = hour.strftime("%H:%M:%S")
 
-        msgCount = int(msgCount)
-        channel = ctx.message.channel
-        botlog = self.client.get_channel(742585460760772709)
-        messages = []
+            msgCount = int(msgCount)+1
+            channel = ctx.message.channel
+            botlog = self.client.get_channel(742585460760772709)
+            messages = []
 
-        async for message in channel.history(limit=msgCount):
-            messages.append(message)
+            async for message in channel.history(limit=msgCount):
+                messages.append(message)
 
-        purgeembed = discord.Embed(
-            title='Purged Messages',
-            description=f'Purged **{msgCount}** Messages in **{channel}**',
-            color=discord.Color.red()
-        )
-        purgeembed.set_footer(text=f'{today} at {hourString}')
+            purgeembed = discord.Embed(
+                title='Purged Messages',
+                description=f'Purged **{msgCount}** Messages in **{channel}**',
+                color=discord.Color.red()
+            )
+            purgeembed.set_footer(text=f'{today} at {hourString}')
 
-        await channel.delete_messages(messages)
-        await ctx.send('Purged Messages')
-        await botlog.send(embed=purgeembed)
+            await channel.delete_messages(messages)
+            await ctx.send('Purged Messages')
+            await botlog.send(embed=purgeembed)
+
+        else:
+            await ctx.send(embed=NoPermsEmbed("Manage Channels"))
 
 def setup(client):
     client.add_cog(Admin(client))
